Improve error marshaling from JS to Ruby
Preserve the message and the name of the error if the runtime provides them.
Also, if possible use Object.assign(...) to copy any additional properties
from the thrown error into the return metadata. Pack the error stack
into the response value together with the error metadata.

Adds :metadata reader to the Ruby error classes so that additional JS
error metadata can be recovered on the Ruby side if desired.